1.完成LED闪灯控制，完成待机与工作状态切换 2.待机功耗控制有待改进 3.电量计算与统计功能还未完成
git-svn-id: https://192.168.1.170:40001/svn/QCD_PowerBank_001@17 c64262a5-5581-5043-b6c8-672724515f9a

+++ sw/tp5602.h
@@ -5,7 +5,7 @@
 #define __TP5602_H__
 
 //电池电压
-u8 GetBattVoltage(void);
+u16 GetBattVoltage(void);
 
 //电池电量
 u8 GetBattCapacity(void);
